Refactor: Apply dependency injection
dependency injection about map utility
