+++ app/src/main/java/campus/tech/kakao/map/utility/MapModule.kt
@@ -1,0 +1,17 @@
+package campus.tech.kakao.map.utility
+
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object MapModule {
+    @Provides
+    @Singleton
+    fun singletonMapUtiliy(): MapUtility{
+        return MapUtility
+    }
+}
